+++ tests/test_parameters.py
@@ -760,7 +760,7 @@ def test_integrated_show_params_option(invoke, create_config):
         ),
         (
             "show-params-cli.theme",
-            "--theme [auto|dark|dracula|light|manpage|monokai|nord|solarized_dark]",
+            "--theme [auto|dark|dracula|light|manpage|monokai|nord|solarized-dark]",
             "click_extra.theme.ThemeOption",
             "click_extra.theme.ThemeChoice",
             "str",